Refactored Cspost component
