+++ pages/blog/[category]/[id].js
@@ -1,14 +1,12 @@
-// import { getAllPostIds, getPostData } from '@/utils/processPosts';
-
 import Date from '@/components/blog/date';
 import Head from '@/components/blog/head';
 import Layout from '@/components/blog/layout';
 import blogStyles from '@/styles/blog/Blog.module.css';
 import styles from '@/styles/blog/Post.module.css';
 import Link from 'next/link';
-import { useState } from 'react';
 import { allPosts } from '@/.contentlayer/generated';
 import { useMDXComponent } from 'next-contentlayer/hooks'
+import { Cspost } from '@/components/blog/cspost';
 
 export function getStaticPaths() {
     const paths = allPosts.map((post) => {
@@ -39,12 +37,16 @@ export async function getStaticProps({ params }) {
     };
 }
 
+const components = {
+    Cspost: ({children}) => <Cspost>{children}</Cspost>,
+}
+
 export default function Post({ postData }) {
     // console.log(postData);
     const MDX = useMDXComponent(postData.body.code)
     return (
         <>
-            <Head 
+            <Head
                 title={postData.title}
                 description={postData.subheader}
                 siteName={`sourish.dev/blog/${postData.category}/${postData.id}`}
@@ -65,76 +67,10 @@ export default function Post({ postData }) {
                     {/* <link rel="stylesheet" href="/blogAssets/css/code.css" /> */}
                     <link rel="stylesheet" href="/blogAssets/css/blog.css" />
                     {/* {postData.cspost ? cspost(postData) : <div className={styles.content} dangerouslySetInnerHTML={{ __html: postData.body.html }} />} */}
-                    <MDX/>
+                    <MDX components={components}/>
                 </div>
             </Layout>
         </>
     )
 }
 
-export function cspost(postData) {
-    const [showNonTech, setShowNonTech] = useState(true);
-
-    return (
-        <>
-            <div className={styles.readingLevel}>
-                <div id="nonTechnicalLink" className={showNonTech ? styles.readingLevelSelected : styles.readingLevelUnselected} onClick={() => { setShowNonTech(true); showNonTechnical() }}>Non-Technical</div>
-                <div id="technicalLink" className={showNonTech ? styles.readingLevelUnselected : styles.readingLevelSelected} onClick={() => { setShowNonTech(false); showTechnical() }}>Technical</div>
-            </div>
-            <div id="nonTechnicalContent" className={styles.content} dangerouslySetInnerHTML={{ __html: postData.nonTechnicalContent }} />
-            <div id="technicalContent" className={styles.content} style={{ display: "none" }} dangerouslySetInnerHTML={{ __html: postData.technicalContent }} />
-        </>
-    )
-}
-
-export function showNonTechnical() {
-    const nonTechnicalContent = document.querySelector("#nonTechnicalContent")
-    const technicalContent = document.querySelector("#technicalContent")
-
-    fadeOut(technicalContent);
-    technicalContent.style.display = "none";
-    nonTechnicalContent.style.display = "block";
-    fadeIn(nonTechnicalContent);
-}
-
-export function showTechnical() {
-    const nonTechnicalContent = document.querySelector("#nonTechnicalContent")
-    const technicalContent = document.querySelector("#technicalContent")
-
-    fadeOut(nonTechnicalContent);
-    nonTechnicalContent.style.display = "none";
-    technicalContent.style.display = "block";
-    fadeIn(technicalContent);
-}
-
-let intervalID;
-
-function fadeOut(fade) {
-    clearInterval(intervalID);
-    var opacity = 1;
-    fade.style.opacity = 1;
-    intervalID = setInterval(function () {
-
-        if (opacity > 0) {
-            opacity = opacity - 0.02;
-            fade.style.opacity = opacity;
-        } else {
-            clearInterval(intervalID);
-        }
-    }, 10);
-}
-
-function fadeIn(fade) {
-    clearInterval(intervalID);
-    var opacity = 0;
-    fade.style.opacity = 0;
-    intervalID = setInterval(function () {
-
-        if (opacity < 1) {
-            opacity = opacity + 0.02;
-            fade.style.opacity = opacity;
-        } else {
-            clearInterval(intervalID);
-        }
-    }, 10);
-}
