+++ posts/nn.mdx
@@ -7,6 +7,8 @@ date:   "2022-07-22"
 category: tutorials
 ---
 
+<Cspost>
+
 Neural Networks are the foundation of all things Deep Learning so it’s pretty natural to be wondering how exactly they work. While they may seem like a black box to most, we find that the box isn’t actually that deep upon opening. With some simple, intuitive diagrams, let’s examine how these fascinating mathematical models learn to play video games, classify animals, and drive cars.
 
 The piece that all neural networks are built on is the neuron. It takes in some inputs and yields an output, very similar to how the neurons in our brain work.
@@ -98,7 +100,7 @@ It's through this process of passing a data point forward, comparing it with the
 
 In the technical section, we’ll see how the back propagation algorithm minimizes this function to obtain the weights that maximize our accuracy. There we'll dive into the linear algebra and calculus for neural networks.
 
-split
+<split></split>
 
 With the help of some linear algebra and calculus, we can build upon our intuition of how neural networks work by understanding the math behind it.
 
@@ -616,3 +618,5 @@ Don't forget that all of this was for just one data point. The entire process ne
 &nbsp; --> */}
 
 I hope the math that goes behind a neural network isn't as daunting for you now. If you have any questions or comments, please don't hesitate to reach out.
+
+</Cspost>
